Fixed some warnings and made the backend more modular
Still a work in progress, but the ultimate goal is to have the backend on a separate module

+++ include/base_mapper.h
@@ -29,8 +29,8 @@ public:
   virtual void prg_write(uint16_t, uint8_t);
   virtual void chr_write(uint16_t, uint8_t);
 
-  template <size_t> void set_prg_map(int, int);
-  template <size_t> void set_chr_map(int, int);
+  template <size_t> void set_prg_map(size_t, int);
+  template <size_t> void set_chr_map(size_t, int);
 
   virtual void scanline_counter();
